[#2]rename: 패키지분리 및 패키지 이름 수정
