+++ src/main/java/com/toyproject/usedauction/domain/category/Category.java
@@ -1,5 +1,6 @@
-package com.toyproject.usedauction.domain.post;
+package com.toyproject.usedauction.domain.category;
 
+import com.toyproject.usedauction.domain.post.Post;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -9,21 +10,17 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 
 @Entity
-public class PostImage {
+public class Category {
 
 	@Id
-	@Column(name = "post_image_id")
+	@Column(name = "category_id")
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
 	@Column(nullable = false)
 	private String name;
 
-	@Column(nullable = false)
-	private String url;
-
 	@ManyToOne
 	@JoinColumn(name = "post_id")
 	private Post post;
-
 }
