+++ src/main/java/com/toyproject/usedauction/domain/postImage/PostImage.java
@@ -1,5 +1,6 @@
-package com.toyproject.usedauction.domain.post;
+package com.toyproject.usedauction.domain.postImage;
 
+import com.toyproject.usedauction.domain.post.Post;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -9,17 +10,21 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 
 @Entity
-public class Category {
+public class PostImage {
 
 	@Id
-	@Column(name = "category_id")
+	@Column(name = "post_image_id")
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
 	@Column(nullable = false)
 	private String name;
 
+	@Column(nullable = false)
+	private String url;
+
 	@ManyToOne
 	@JoinColumn(name = "post_id")
 	private Post post;
+
 }
